prelim gdd model and simulation test
File "gdd.stan" is a Bayesian model of budburst wherein plants
accumulate growth daily when temperatures are above a
minimum (Tmin). They then bud on the day that accumulated growth
reaches a threshold. The threshold is assumed to be normally
distributed with mean = mu_gdd and sd = sigma_gdd. File
"simulation_budburst_analysis_GDD.R" simulates the model and tests
"gdd.stan" by fitting it to the simulated data.